refactor: one route for rendering static file
