+++ app.js
@@ -6,10 +6,9 @@ const connectDB = require("./db/connect");
 const app = express();
 
 // Routes Import
-const register = require("./routes/register");
-const login = require("./routes/login");
-const dashboard = require("./routes/dashboard");
-const user = require("./routes/user");
+const pages = require("./routes/page");
+const users = require("./routes/user");
+const posts = require("./routes/post");
 
 //ENV
 require("dotenv").config();
@@ -20,10 +19,9 @@ app.use(express.static("./public"));
 app.use(express.json());
 
 // Routes
-app.use("/register", register);
-app.use("/login", login);
-app.use("/dashboard", dashboard);
-app.use("/user", user);
+app.use(pages);
+app.use("/user", users);
+app.use("/post", posts);
 
 // Homepage (Nanti)
 app.get("/", (req, res) => {
