--- controllers/dashboardController.js
@@ -1,9 +1,0 @@
-const path = require("path");
-
-const getDashboardPage = async () => {
-  res.sendFile(path.join(__dirname, "..", "public", "html", "dashboard.html"));
-};
-
-module.exports = {
-  getDashboardPage,
-};
