--- routes/dashboard.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const { getDashboardPage } = require("../controllers/dashboardController");
-
-router.route("/").get(getDashboardPage);
-
-module.exports = router;
