--- routes/register.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const { getRegisterPage } = require("../controllers/registerController");
-
-router.route("/").get(getRegisterPage);
-
-module.exports = router;
